+++ routes/teachers.js
@@ -1,6 +1,7 @@
 const express = require("express")
 const router = express.Router()
 const TeachersController = require("../controller/teachers")
+const CalendarsController = require("../controller/calendars")
 const {
   validateRegisterCourse,
   validateEditTeacherInfo,
@@ -30,17 +31,16 @@ router.delete(
   validateDeleteCourse,
   TeachersController.deleteCourse
 )
-
 router.get(
   "/calendar",
   validateGetCalendarInfo,
-  TeachersController.getCalendarInfo
+  CalendarsController.getCalendarInfo
 )
-router.post("/calendar", validateAddCalendar, TeachersController.addCalendar)
+router.post("/calendar", validateAddCalendar, CalendarsController.addCalendar)
 router.delete(
   "/calendar",
   validateDeleteCalendar,
-  TeachersController.deleteCalendar
+  CalendarsController.deleteCalendar
 )
 
 module.exports = router
